Move components to separate dir
